Pass the Resolver to the onRequestCancel callback
This way it doesn't have to capture the original resolver, which avoids
a retain cycle in the event that the promise is never resolved or
cancelled.

This also means we can't invoke `onRequestCancel` callbacks when the
`PromiseBox` deinits anymore, but that should have only actually done
anything if a `Resolver` was thrown away without ever being used to
resolve the `Promise`. We were treating that as an implicit cancellation
but now we don't.