Relax guarantees about promise context resolution
Don't hop onto the context when resolving promises if we're not calling
a registered callback handler in order to do so. This means e.g. `.then`
no longer hops onto the context if it's passing through a rejection or a
cancellation.

This was never something we documented, and it really only affects
callbacks registered with `.immediate`.