Ensure all callback registration methods return a new Promise
This means `.catch`, `.always`, and `.onCancel` all return a new
`Promise`, and `-then:`, `-catch:`, `-inspect:`, and `-whenCancelled:`
all return a new `TWLPromise`.

The reason for this is so chained callbacks always guarantee that the
previous callback finishes executing before the next one starts. We're
paying a small cost to do this, and the behavior is indistinguishable
when the chained callbacks all use a serial queue (such as `.main`), but
it will prevent potentially surprising behavior when using concurrent
contexts (e.g. `.utility`). With the old behavior the only way to get
this new behavior was to use `.always` and manually return a new
`Promise`. With the new behavior you can recover the old simply by
chaining multiple callbacks off of the same promise manually (though you
still pay the cost for creating the new promises).