Improve the behavior of delay using an OperationQueue context
Previously we were only enqueuing an operation on the operation queue
when the delay was finished. This meant that if the queue was backed up,
our operation would have the lowest priority. Instead, enqueue an
operation immediately that simply isn't ready yet, and then mark it as
ready when the delay elapses. This ensures our operation will run sooner
if the queue is backed up.

Also fix the documentation as it made the wrong claim about our
operation queue behavior.